refactor: migrate the pop-up-box component and its logic
