+++ src/app/components/pokemon-page-client/pokemon-page-client.tsx
@@ -1,10 +1,12 @@
 'use client';
 
 import { useState } from 'react';
+import { useSelector } from 'react-redux';
 
 import PokemonList from '../../components/pokemon-list/pokemon-list';
 import SearchField from '../../components/search-field/search-field';
-import { Pokemon } from '../../utils/types';
+import { Pokemon, RootState } from '../../utils/types';
+import PopUpBox from '../pop-up-box/pop-up-box';
 
 interface PokemonPageClientProps {
   initialPokemons: Pokemon[];
@@ -15,6 +17,9 @@ export default function PokemonPageClient({
 }: PokemonPageClientProps) {
   const [filteredPokemons, setFilteredPokemons] =
     useState<Pokemon[]>(initialPokemons);
+  const selectedCount = useSelector(
+    (state: RootState) => state.checkbox.selectedCount
+  );
 
   return (
     <>
@@ -23,6 +28,7 @@ export default function PokemonPageClient({
         onSearch={(filtered) => setFilteredPokemons(filtered)}
       />
       <PokemonList pokemonList={filteredPokemons} />
+      {selectedCount > 0 && <PopUpBox selectedCount={selectedCount} />}
     </>
   );
 }
